This commit introduces substantial refactoring and restructuring of the codebase to improve modularity and add new functionalities.
- Added `build.ts`, `parse.ts`, and `read.ts` files under `src/` to modularize and separate main functionalities of building, parsing, and reading EML content.
- Implemented `createBoundary` and `completeBoundary` functions in `build.ts` to handle email boundary creation and processing respectively.
- Enhanced charset handling through `getCharset` function in `charset.ts` and defined `DEFAULT_CHARSET` as 'utf-8' in the new `constants.ts`.
- `header.ts` includes email-related utilities, such as `toEmailAddress`, `getEmailAddress`, and functions for encoding/decoding headers.
- Updated `interface.ts` with precise types for EML content transformation.
- Refactored `index.ts` to export necessary functionalities and types from newly created modules.
- Implemented a comprehensive test suite spread across multiple files to ensure the functionality and stability of newly introduced modules, including `build.test.ts`, `parse.test.ts`, `read.test.ts`, and `header.test.ts`.

These changes enhance code clarity and maintainability, ensuring robust handling of email content in different formats.
