+++ src/constants.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_CHARSET = 'utf-8';
+
