+++ test/constants.test.ts
@@ -1,0 +1,7 @@
+import { DEFAULT_CHARSET } from '../src/constants';
+
+describe('Constants', () => {
+    it('DEFAULT_CHARSET should be utf-8', () => {
+        expect(DEFAULT_CHARSET).toBe('utf-8');
+    });
+});
